Rename shape_adapter to graph_element_adapter
