+++ patterns/adapter/text_graphics/shapes/graph_element_adapter.dart
